Changed for loop from "for ... in" to classic for
The current for...in loop works fine in most cases, but breaks in IE in
certain cases.

However, if a developer happens to extend the Array object (in my case
to add an "indexOf" function), the for...in loop causes problems in
Internet Explorer 8 and below. IE will attempt to iterate over the
function as an element and throws an error.  Changing it to a classic
for loop fixes that problem.